+++ lib/dataLayer.tsx
@@ -179,6 +179,7 @@ async function _getFormByID(formID: string): Promise<PublicFormSchemaProperties 
           formID,
           ...records[0].formConfig.form,
           publishingStatus: records[0].formConfig.publishingStatus,
+          displayAlphaBanner: records[0].formConfig.displayAlphaBanner ?? true,
         };
       }
       return null;
@@ -203,6 +204,7 @@ async function _getFormByID(formID: string): Promise<PublicFormSchemaProperties 
             formID,
             ...records[0].formConfig?.form,
             publishingStatus: records[0].formConfig.publishingStatus,
+            displayAlphaBanner: records[0].formConfig.displayAlphaBanner ?? true,
           };
         }
         return null;
@@ -229,6 +231,7 @@ async function _getFormByStatus(
                 formID: record.formID,
                 ...record.formConfig.form,
                 publishingStatus: record.formConfig.publishingStatus,
+                displayAlphaBanner: record.formConfig.displayAlphaBanner ?? true,
               };
             }
           })
@@ -260,6 +263,7 @@ async function _getFormByStatus(
                   formID: record.formID,
                   ...record.formConfig?.form,
                   publishingStatus: record.formConfig?.publishingStatus,
+                  displayAlphaBanner: record.formConfig.displayAlphaBanner ?? true,
                 };
               }
             })
@@ -316,7 +320,9 @@ export function extractFormData(submission: Submission): Array<string> {
   const formOrigin = submission.form;
   const dataCollector: Array<string> = [];
   formOrigin.layout.map((qID) => {
-    const question = formOrigin.elements.find((element: FormElement) => element.id === qID);
+    const question = formOrigin.elements.find(
+      (element: FormElement) => element.id === parseInt(qID)
+    );
     if (question) {
       handleType(question, formResponses[question.id], dataCollector);
     }
@@ -427,26 +433,26 @@ function _handleFormDataType(element: FormElement, value: Response, formData: Fo
     case "textField":
     case "textArea":
       // string
-      _handleFormDataText(element.id, value as string, formData);
+      _handleFormDataText(element.id.toString(), value as string, formData);
       break;
 
     case "dropdown":
     case "radio":
       value instanceof Object
-        ? _handleFormDataText(element.id, "", formData)
-        : _handleFormDataText(element.id, value, formData);
+        ? _handleFormDataText(element.id.toString(), "", formData)
+        : _handleFormDataText(element.id.toString(), value, formData);
       break;
 
     case "checkbox":
     case "dynamicRow":
       // array of strings
       Array.isArray(value)
-        ? _handleFormDataArray(element.id, value as Array<string>, formData)
-        : _handleFormDataText(element.id, "", formData);
+        ? _handleFormDataArray(element.id.toString(), value as Array<string>, formData)
+        : _handleFormDataText(element.id.toString(), "", formData);
       break;
     case "fileInput":
       // file input
-      _handleFormDataFileInput(element.id, value as FileInputResponse, formData);
+      _handleFormDataFileInput(element.id.toString(), value as FileInputResponse, formData);
       break;
   }
 }
@@ -469,7 +475,7 @@ async function _submitToAPI(values: Responses, formikBag: FormikBag<DynamicFormP
   const formDataObject = _buildFormDataObject(formConfig, values);
 
   //making a post request to the submit API
-  await axios({
+  return await axios({
     url: "/api/submit",
     method: "POST",
     headers: {